src/2. BigData/2. Models/3. advanced/pyproject.toml
[project]
name = "iowa-timeseries-advanced"
version = "0.1.0"
description = "Advanced AutoGluon time series forecasting with evaluation, caching, and API-ready outputs"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0,<2.0.0",
    "torch>=2.2.0,<2.3.0",
    "autogluon.timeseries>=1.0.0",
    "matplotlib>=3.7.0",
    "google-cloud-bigquery>=3.0.0",
    "db-dtypes>=1.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
]
